// models/charts/machineModelChartDTO.ts
export interface MachineModelChartDTO{
    brandName : string;
    total     : number;
    clientName: String;
    marca     : String;
    parqueMaquinas : number;
    type : string;
    location:string;
    concess:string;
    ano: string;
    cotton: string;
    cvTrator: string;
    entreLinhas: string;
    metrosCubicos: string;
    numLinhas: string;
    pesColheitadeira: string;
    tamanhoBarra: string;
    tipoEquipamento: string;
    tipoPlataforma: string;
    detailTable:String;


}